fix: make utils folder for globalfile and rename exception-global.filter
