+++ src/utils/validate-exceptions.ts
